Extract clean up worker out
